Switched to bitfields for 0x79 and 0x6C, Changed various funcs to use floats, Matched chmole func, added maps enum, added comments

+++ include/ch/exploder.h
@@ -5,6 +5,6 @@
 
 void _chexploder_entrypoint_3(void*, void*, u32);
 void _chexploder_entrypoint_13(s32, u32);
-void _chexploder_entrypoint_15(s32, u32, u32);
+void _chexploder_entrypoint_15(s32, u32, f32);
 
 #endif
