+++ include/gs/propctrl.h
@@ -2,6 +2,6 @@
 #define __GS_PROPCTRL_H__
 #include "common.h"
 
-void _gspropctrl_entrypoint_11(Prop*, u32*);
+void _gspropctrl_entrypoint_11(Prop*, f32*);
 
 #endif
